Rename html/{->html}{label,title}.ts
