+++ bokehjs/src/lib/models/annotations/html/html_label.ts
